move resize and rotate to new "transform" module
